move dataloader utils from "mixin" to utils
